+++ Editor/UI/ProjectAuditorWindow.cs
@@ -146,7 +146,7 @@ namespace Unity.ProjectAuditor.Editor.UI
             {
                 var desc = viewDescriptors[i];
                 var layout = m_ProjectAuditor.GetLayout(desc.category);
-                var isSupported = layout != null || desc.category == IssueCategory.None;
+                var isSupported = layout != null;
 
                 m_ViewDropdownItems[i] = new Utility.DropdownItem
                 {
@@ -169,6 +169,8 @@ namespace Unity.ProjectAuditor.Editor.UI
 
                 m_Views[i] = view;
             }
+
+            SummaryView.SetReport(m_ProjectReport);
 
             var shaderVariantsWindow = AnalysisWindow.FindOpenWindow<ShaderVariantsWindow>();
             if (shaderVariantsWindow != null)
@@ -233,6 +235,17 @@ namespace Unity.ProjectAuditor.Editor.UI
         [InitializeOnLoadMethod]
         static void OnLoad()
         {
+            ViewDescriptor.Register(new ViewDescriptor
+            {
+                category = IssueCategory.MetaData,
+                name = "Summary",
+                menuOrder = -1,
+                showActions = false,
+                showFilters = false,
+                showInfoPanel = true,
+                viewType = typeof(SummaryView),
+                analyticsEvent = (int)ProjectAuditorAnalytics.UIButton.Summary
+            });
             ViewDescriptor.Register(new ViewDescriptor
             {
                 category = IssueCategory.Assets,
@@ -382,6 +395,8 @@ namespace Unity.ProjectAuditor.Editor.UI
                     view.Clear();
             }
 
+            SummaryView.SetReport(m_ProjectReport);
+
             if (m_ShaderVariantsWindow != null)
             {
                 m_ShaderVariantsWindow.Clear();
@@ -446,7 +461,6 @@ namespace Unity.ProjectAuditor.Editor.UI
             var newIssues = new List<ProjectIssue>();
             auditor.Audit(issue =>
             {
-
                 newIssues.Add(issue);
                 m_ProjectReport.AddIssue(issue);
             },
@@ -1122,7 +1136,7 @@ To Analyze the project, click on Analyze.
 Once the project is analyzed, the tool displays a list of issues of a specific kind. Initially, code-related issues will be shown.
 To switch type of issues, for example from code to settings-related issues, use the 'View' dropdown and select Settings.
 In addition, it is possible to filter issues by area (CPU/Memory/etc...), by string or by other search criteria."
-                );
+            );
 
             public static readonly GUIContent ShaderVariantsButton = new GUIContent("Variants", "Inspect Shader Variants");
         }