+++ Tests/Editor/CustomAuditorTest.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using NUnit.Framework;
@@ -13,7 +13,7 @@ namespace UnityEditor.ProjectAuditor.EditorTests
         private readonly IssueLayout m_Layout = new IssueLayout
         {
             category = IssueCategory.Code,
-            properties = new []
+            properties = new[]
             {
                 new PropertyDefinition {type = PropertyType.Description }
             }