+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* Added Build Report view (experimental)
 * Added Generic types instantiation analysis (experimental)
+* Added Summary view
 
 ## [0.5.0-preview] - 2021-03-11
 * Fixed reporting of issues affecting multiple areas